Add and Update Assets, Services, and UI Components
- Added new images: `gallery.png` and updated `splash.gif`.
- Updated `app_assets.dart` to include new asset paths for `splash.gif` and `gallery.png`.
- Introduced a new primary color in `app_colors.dart`.
- Enhanced `data_service.dart` with new methods for deleting documents and images, and checking image existence.
- Implemented new functionalities in `firestore_service.dart` to support document and image deletion, and image existence checks.
- Updated `app_router.dart` to include new routes for removing items from the store and fixed the splash view route.
- Modified `custom_number_field.dart` to parse input as integers.
- Added a new widget `show_delete_confirmation_dialog.dart` for displaying delete confirmation dialogs.
- Implemented new repository methods in `admin_repo_impl.dart` for deleting phone data and associated images.
- Added new cubit and state classes in `remove_from_store_cubit.dart` and `remove_from_store_state.dart` for managing the removal of items from the store.
- Created new views and widgets for the remove from store feature, including `remove_from_store_view.dart`, `phone_data_element.dart`, and `remove_from_store_view_body.dart`.
- Updated `add_in_store_view_body_bloc_consumer.dart` to adjust image dimensions and spacing.
- Modified `admin_view_body.dart` to update the route for removing items from the store.
- Updated `onboarding_static_data.dart` to use the new primary color.
- Enhanced the splash view with a sliding text animation and updated the background color in `splash_view.dart`.
- Added a new widget `sliding_image.dart` for the sliding text animation in the splash view.
- Updated `splash_view_body.dart` to include the new sliding text animation and adjusted the splash screen layout.

+++ lib/core/widgets/custom_number_field.dart
@@ -32,7 +32,7 @@ class CustomNumberField extends StatelessWidget {
         if (value == null || value.isEmpty) {
           return 'هذا الحقل مطلوب';
         }
-        final number = double.tryParse(value);
+        final number = int.tryParse(value);
         if (number == null) {
           return 'ادخل رقم صحيح';
         }
